Introduce the `with-untramped-file'
This macro calls `dctrl-untramp-file' to generate the necessary
commands to transfer FILE onto the host of the controller.  It appends
these commands with the commands generated by BODY.  The FILE variable
is set with the file path on the host of the controller.  It allows
simple construction like the following:

  (with-untramped-file file
    (dctrl-run-process (list \"fastboot\" \"flash\" \"boot\" file)))"

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>